+++ src/apis/spaces/private-spaces/dto/create-private-space.input.ts
