+++ src/apis/spaces/private-spaces/entities/private-space.entity.ts
