+++ src/apis/spaces/private-spaces/private-spaces.service.ts
