+++ src/apis/spaces/stores/dto/create.store.dto.ts
