+++ src/apis/spaces/stores/dto/search.store.dto.ts
