+++ src/apis/spaces/stores/entities/store.entity.ts
