+++ src/apis/spaces/stores/stores.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { FranchisesModule } from '../franchises/franchises.module';
+import { FranchisesModule } from '../../franchises/franchises.module';
 import { Store } from './entities/store.entity';
 import { StoresResolver } from './stores.resolver';
 import { StoresService } from './stores.service';
