+++ src/apis/spaces/stores/stores.service.ts
