+++ src/apis/tumbler-records/entities/tumbler-record.entity.ts
@@ -1,5 +1,5 @@
 import { Field, InputType, Int, ObjectType } from '@nestjs/graphql';
-import { Store } from 'src/apis/stores/entities/store.entity';
+import { Store } from 'src/apis/spaces/stores/entities/store.entity';
 import { User } from 'src/apis/users/entities/user.entity';
 import { CommonEntity } from 'src/commons/entities/common.entity';
 import { Column, Entity, JoinColumn, ManyToOne } from 'typeorm';
